Create RandomTree.md
Code for each tree in this random forest.

All I really have to do left is prove that such a partition as commented out in line 79 exists (don't want to write code to find something if it doesn't always exist). I want to prove that, given a sorted set of randomly selected integers, there exists a split of that set into L1 and L2 meeting the following conditions: WLOG every element in L1 < every element in L2, at least one of L1 or L2 has strictly more even or odd numbers than the other set and that other set cannot have less elements of the other parity than the first set, and at least one of L1 or L2 has a strictly larger proportion of either even or odd numbers than the other set and that other set cannot have a smaller proportion of numbers of the other parity than in the first set.